+++ include/aedis/ssl/connection.hpp
@@ -15,6 +15,102 @@
 
 namespace aedis::ssl {
 
+namespace detail
+{
+
+#include <boost/asio/yield.hpp>
+
+template <class Stream>
+struct handshake_op {
+   Stream* stream;
+   aedis::detail::conn_timer_t<typename Stream::executor_type>* timer;
+   boost::asio::coroutine coro{};
+
+   template <class Self>
+   void operator()( Self& self
+                  , std::array<std::size_t, 2> order = {}
+                  , boost::system::error_code ec1 = {}
+                  , boost::system::error_code ec2 = {})
+   {
+      reenter (coro)
+      {
+         yield
+         boost::asio::experimental::make_parallel_group(
+            [this](auto token)
+            {
+               return stream->async_handshake(boost::asio::ssl::stream_base::client, token);
+            },
+            [this](auto token) { return timer->async_wait(token);}
+         ).async_wait(
+            boost::asio::experimental::wait_for_one(),
+            std::move(self));
+
+         switch (order[0]) {
+            case 0: self.complete(ec1); return;
+            case 1:
+            {
+               BOOST_ASSERT_MSG(!ec2, "handshake_op: Incompatible state.");
+               self.complete(error::ssl_handshake_timeout);
+               return;
+            }
+
+            default: BOOST_ASSERT(false);
+         }
+      }
+   }
+};
+
+template <
+   class Stream,
+   class CompletionToken
+   >
+auto async_handshake(
+      Stream& stream,
+      aedis::detail::conn_timer_t<typename Stream::executor_type>& timer,
+      CompletionToken&& token)
+{
+   return boost::asio::async_compose
+      < CompletionToken
+      , void(boost::system::error_code)
+      >(handshake_op<Stream>{&stream, &timer}, token, stream, timer);
+}
+
+template <class Conn>
+struct ssl_connect_with_timeout_op {
+   Conn* conn = nullptr;
+   boost::asio::coroutine coro{};
+
+   template <class Self>
+   void operator()( Self& self
+                  , boost::system::error_code ec = {}
+                  , boost::asio::ip::tcp::endpoint const& = {})
+   {
+      reenter (coro)
+      {
+         conn->ping_timer_.expires_after(conn->get_config().connect_timeout);
+
+         yield
+         aedis::detail::async_connect(
+            conn->lowest_layer(), conn->ping_timer_, conn->endpoints_, std::move(self));
+
+         if (ec) {
+            self.complete(ec);
+            return;
+         }
+
+         conn->ping_timer_.expires_after(conn->get_config().handshake_timeout);
+
+         yield
+         async_handshake(conn->next_layer(), conn->ping_timer_, std::move(self));
+         self.complete(ec);
+      }
+   }
+};
+
+#include <boost/asio/unyield.hpp>
+
+} // detail
+ 
 template <class>
 class connection;
 
@@ -34,6 +130,7 @@ public:
    using executor_type = typename next_layer_type::executor_type;
 
    using base_type = connection_base<executor_type, connection<boost::asio::ssl::stream<AsyncReadWriteStream>>>;
+   using this_type = connection<next_layer_type>;
 
    /** @brief Connection configuration parameters.
     */
@@ -43,6 +140,9 @@ public:
 
       /// Timeout of the connect operation.
       std::chrono::milliseconds connect_timeout = std::chrono::seconds{10};
+
+      /// Timeout of the ssl handshake operation.
+      std::chrono::milliseconds handshake_timeout = std::chrono::seconds{10};
 
       /// Time interval of ping operations.
       std::chrono::milliseconds ping_interval = std::chrono::seconds{1};
@@ -86,19 +186,18 @@ public:
    auto& lowest_layer() noexcept { return stream_.lowest_layer(); }
    auto is_open() const noexcept { return stream_.next_layer().is_open(); }
 
-   template <
-      class EndpointSequence,
-      class CompletionToken
-      >
-   auto async_connect(
-         detail::conn_timer_t<executor_type>& timer,
-         EndpointSequence ep,
-         CompletionToken&& token)
+   template <class CompletionToken>
+   auto async_connect(CompletionToken&& token)
    {
-      return detail::async_connect(lowest_layer(), timer, ep, std::move(token));
+      return boost::asio::async_compose
+         < CompletionToken
+         , void(boost::system::error_code)
+         >(detail::ssl_connect_with_timeout_op<this_type>{this}, token, stream_);
    }
 
 private:
+   template <class> friend struct detail::ssl_connect_with_timeout_op;
+
    config cfg_;
    executor_type ex_;
    next_layer_type stream_;